pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdlc-orchestrator"
version = "0.1.0"
description = "Autonomous SDLC orchestration system with human approval gates"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
sdlc = "sdlc_orchestrator.cli:cli"

[tool.setuptools.packages.find]
include = ["sdlc_orchestrator*"]

[tool.setuptools.package-data]
sdlc_orchestrator = ["skills/*.md", "templates/*"]
